App icon redesign + V4 home screen + shift premium decimal support
- Icon: three-S triskelion pattern replacing calendar icon, with amber/teal/violet
  S-curves at 120° rotation representing SimpleShiftScheduler and shift rotation
- Home: NewHomeScreenV4 with unified card language (V2CardShape throughout),
  circle-badge hero pattern, consistent 16dp spacing, decoupled stat cards
- Salary: shift premium changed from Int to Double across all layers (SalaryConfig,
  SalaryBreakdown, salary_calculator, SettingsRepository, ViewModel, Screen)
- Input: premium text fields use local text state to allow partial input like "5."
  without bouncing back to "0"
- Tests: updated SalaryCalculatorTest and SettingsRepositoryTest for Double values,
  added decimal premium calculation test case
- Docs: updated memory-bank/architecture.md and app-design-document.md

+++ app/src/main/java/com/simpleshift/scheduler/domain/model/SalaryBreakdown.kt
@@ -5,5 +5,5 @@ import java.time.YearMonth
 data class SalaryBreakdown(
     val month: YearMonth,
     val shiftCounts: Map<ShiftType, Int>,
-    val shiftPremiumTotal: Int
+    val shiftPremiumTotal: Double
 )
