+++ app/src/main/java/com/simpleshift/scheduler/domain/model/SalaryConfig.kt
@@ -1,5 +1,5 @@
 package com.simpleshift.scheduler.domain.model
 
 data class SalaryConfig(
-    val shiftPremiums: Map<ShiftType, Int> = emptyMap()
+    val shiftPremiums: Map<ShiftType, Double> = emptyMap()
 )
